backend/requirements.txt
aniso8601==9.0.1
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Cython==0.29.21
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
gensim==3.6.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
mysql-connector-python==8.0.23
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pathy==0.5.2
preshed==3.0.5
protobuf==3.16.0
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-Levenshtein==0.12.2
pytorch-crf==0.7.2
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
threadpoolctl==2.1.0
torch==1.8.1
torchtext==0.6.0
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==1.0.1
